koa-blog-api: add more user specific checks
* extract current user from auth header's token
* and check if it matches author of content to be edited/deleted
* send errors as response body with error status
* move jwt to utils

+++ index.js
@@ -8,7 +8,7 @@ const mongo = require('./utils/mongo'),
 const bodyParser = require('koa-bodyparser')
 app.use(bodyParser())
 
-const { errorHandler, jwt, issue }  = require("./jwt");
+const { errorHandler, getUserFromToken, jwt, issue }  = require("./utils/jwt");
 
 // authentication
 require('./utils/passport')
@@ -41,7 +41,8 @@ public
   .post("/login", bodyParser(), async ctx => {
     await passport.authenticate('local', (error, user) => {
       if (user == false || error) {
-        throw Error(error || 'Authorization error')
+        ctx.body = { message : error || 'Authorization error'}
+        ctx.status = 401
       } else {
         const payload = {
           id: user.id,
@@ -61,15 +62,20 @@ public
         password: Joi.string().max(100),
       },
       type: 'json',
+      continueOnError: true
     },
     handler: async ctx => {
+      if (ctx.invalid && typeof ctx.invalid !== 'undefined') {
+        ctx.status = ctx.invalid.body.status
+        return ctx.body = { message: ctx.invalid.body.msg }
+      }
       const user = await new User(ctx.request.body);
       try {
         await user.save();
         ctx.body = await user;
       } catch(err) {
         ctx.status = 403
-        ctx.body = { message: err}
+        ctx.body = { message: err.message}
       }
     }
   })
@@ -81,12 +87,24 @@ secured
     ctx.body = await post;
   })
   .delete('/posts/:id', async ctx => {
-    await Post.deleteOne({ _id: ObjectId(ctx.params.id) })
-    ctx.body = { message: "Successfully deleted" };
+    const isAuthor = await checkPostAuthor(ctx.headers.authorization, ctx.params.id)
+    if (isAuthor) {
+      await Post.deleteOne({ _id: ObjectId(ctx.params.id) })
+      return ctx.body = { message: "Successfully deleted" };
+    } else {
+      ctx.body = { message: "You are not the author" }
+      ctx.status = 400
+    }
   })
   .put('/posts/:id', bodyParser(), async ctx => {
-    await Post.updateOne({ _id: ObjectId(ctx.params.id) }, {$set: {"name": ctx.request.body.name}})
-    ctx.body = { message: "Successfully updated" };
+    const isAuthor = await checkPostAuthor(ctx.headers.authorization, ctx.params.id)
+    if (isAuthor) {
+      await Post.updateOne({ _id: ObjectId(ctx.params.id) }, {$set: {"name": ctx.request.body.name}})
+      ctx.body = { message: "Successfully updated" };
+    } else {
+      ctx.status = 400
+      ctx.body = { message: "You are not the author" }
+    }
   })
   .post('/posts/:id/comments/', bodyParser(), async ctx => {
     const commentId = ObjectId();
@@ -98,17 +116,40 @@ secured
     ctx.body = commentsDocument.comments[0];
   })
   .delete('/posts/:id/comments/:cid/', async ctx => {
-    await Post.updateOne({ _id: ObjectId(ctx.params.id) },
-      {$pull: {comments: {_id: ObjectId(ctx.params.cid)}}})
-    ctx.body = { message: "Successfully deleted comment" };
+    const isAuthor = await checkCommentAuthor(ctx.headers.authorization, ctx.params.id, ctx.params.cid)
+    if (isAuthor) {
+      await Post.updateOne({ _id: ObjectId(ctx.params.id) },
+        {$pull: {comments: {_id: ObjectId(ctx.params.cid)}}})
+      ctx.body = { message: "Successfully deleted comment" };
+    } else {
+      ctx.status = 400
+      ctx.body = { message: "You are not the author" }
+    }
   })
   .put('/posts/:id/comments/:cid/', bodyParser(), async ctx => {
-    await Post.updateOne(
-      { _id: ObjectId(ctx.params.id), "comments._id": ObjectId(ctx.params.cid) },
-      { $set: { "comments.$.name" : ctx.request.body.name } }
-   )
-   ctx.body = { message: "Successfully updated comment" };
+    const isAuthor = await checkCommentAuthor(ctx.headers.authorization, ctx.params.id, ctx.params.cid)
+    if (isAuthor) {
+      await Post.updateOne(
+        { _id: ObjectId(ctx.params.id), "comments._id": ObjectId(ctx.params.cid) },
+        { $set: { "comments.$.name" : ctx.request.body.name } }
+      )
+      ctx.body = { message: "Successfully updated comment" };
+    } else {
+      ctx.status = 400
+      ctx.body = { message: "You are not the author" }
+    }
   })
 
+const checkPostAuthor = (token, pid) => {
+  const user = getUserFromToken(token)
+  return Post.findById(ObjectId(pid))
+    .then(post => post.author === user.username)
+}
+
+const checkCommentAuthor = (token, pid, cid) => {
+  const user = getCurrentUser(token)
+  return Post.findOne({_id: pid},{comments: {$elemMatch: {_id: cid}}})
+    .then(post => post.comments[0].author === user.username)
+}
 const port = process.env.PORT || 3000;
 app.listen(port, () => console.log('Server listening on', port))
